rename server/controllers/auth/signed_url_action.ts to server/controllers/auth/signed_url.ts
