+++ server/controllers/auth/signed_url.ts
